editor/libeditor/html/tests/test_bug612128.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=612128
-->
<head>
  <title>Test for Bug 612128</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=612128">Mozilla Bug 612128</a>
<p id="display"></p>
<div id="content">
<input>
<div contenteditable></div>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 612128 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  document.querySelector("input").focus();
  var threw = false;
  try {
    document.execCommand("inserthtml", null, "<span>f" + "oo</span>");
  } catch (e) {
    threw = true;
  }
  ok(threw, "The inserthtml command should fail");
  is(document.querySelectorAll("span").length, 0, "No span element should be injected inside the page");
  is(document.body.innerHTML.indexOf("f" + "oo"), -1, "No text should be injected inside the page");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
